+++ lib/ui/page/home/tab/contacts/view.dart
@@ -449,15 +449,22 @@ class ContactsTabView extends StatelessWidget {
                 key: const Key('UnfavoriteContactButton'),
                 label: 'btn_delete_from_favorites'.l10n,
                 onPressed: () => c.unfavoriteContact(contact.contact.value.id),
+                trailing: const Icon(Icons.star_border),
               )
             : ContextMenuButton(
                 key: const Key('FavoriteContactButton'),
                 label: 'btn_add_to_favorites'.l10n,
                 onPressed: () => c.favoriteContact(contact.contact.value.id),
+                trailing: const Icon(Icons.star),
               ),
         ContextMenuButton(
-          label: 'btn_delete_from_contacts'.l10n,
+          label: 'btn_delete'.l10n,
           onPressed: () => c.deleteFromContacts(contact.contact.value),
+          trailing: const Icon(Icons.delete),
+        ),
+        ContextMenuButton(
+          label: 'btn_select'.l10n,
+          trailing: const Icon(Icons.select_all),
         ),
       ],
       subtitle: [