+++ CHANGELOG.md
@@ -56,6 +56,7 @@ All user visible changes to this project will be documented in this file. This p
     - Profile page:
         - Redesigned profile page. ([#244], [#257], [#243])
     - Redesigned desktop context menu. ([#245])
+    - Redesigned mobile context menu. ([#305], [#295])
     - Auth page:
         - Redesigned login modal. ([#249])
     - User page:
@@ -146,7 +147,9 @@ All user visible changes to this project will be documented in this file. This p
 [#288]: /../../pull/288
 [#292]: /../../issues/292
 [#293]: /../../pull/293
+[#295]: /../../issues/295
 [#300]: /../../pull/300
+[#305]: /../../pull/305
 [#310]: /../../pull/310